move css modules next to components
