+++ components/RightPanel.module.css
@@ -1,0 +1,44 @@
+.rightPanel {
+  position: absolute;
+  top: 60px;
+  right: 0;
+  width: 340px;
+  background: var(--panel-bg);
+  color: var(--text-color);
+  padding: 12px;
+  display: flex;
+  flex-direction: column;
+  height: calc(100% - 60px);
+  font-size: 16px;
+}
+
+.emotions {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 4px;
+}
+
+.emotions label {
+  display: flex;
+  align-items: center;
+  gap: 6px;
+  padding: 4px 6px;
+  border-radius: 4px;
+  color: #fff;
+  font-size: 14px;
+  cursor: pointer;
+}
+
+.feed {
+  flex: 1;
+  list-style: none;
+  padding: 0;
+  margin: 12px 0;
+  overflow-y: auto;
+}
+
+.pagination {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
