+++ components/TopBar.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import styles from '../styles/TopBar.module.css';
+import styles from './TopBar.module.css';
 
 export default function TopBar({
   view,
